requirements.txt
streamlit
pandas
openpyxl
requests
beautifulsoup4
lxml
python-docx
PyPDF2
matplotlib
seaborn
openai>=1.0.0

